requirements.txt
# Required Python packages for Robot Mower Advanced

# Core dependencies
numpy>=1.22.0
PyYAML>=6.0
RPi.GPIO>=0.7.0
pyserial>=3.5
smbus2>=0.4.2
pigpio>=1.78

# Web interface
Flask>=2.1.0
Flask-SocketIO>=5.2.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
eventlet>=0.33.0
gunicorn>=20.1.0

# Computer vision and processing
opencv-python-headless>=4.6.0
Pillow>=9.0.0

# Navigation and path planning
matplotlib>=3.5.0
scipy>=1.8.0
shapely>=1.8.0

# Networking and communication 
requests>=2.27.0
paho-mqtt>=1.6.0

# Documentation
Sphinx>=4.5.0

# Testing
pytest>=7.0.0
pytest-mock>=3.7.0
